// src/utils/shared-types.ts
export type reportType = {
  periodOfTime?: string;

  gender?: string;
  age?: string;
  genderFreeField?: string;
  organizationType?: string[];
  numberOfEmployees?: string;
  organizationTypeFreeField?: string;
  happenedOnlineFreeField?: string;
  PeriodOfTimeFreeField?: string;
  locationOnline?: string;
  identity?: string;
  dateRangeState?: any;
  valueDate?: any;

  location?: any;
  _id?: string;

  happenedOverALongPeriodOfTime?: string;

  manifestationOfDiscrimination?: string;
  manifestationOfDiscriminationFreeField?: string;
  otherFormOfDisc?: string;
  otherFormOfDiscFreeField?: string;
  otherFormOfDiscYes?: string;
  haveYouReported?: string;
  haveYouReportedYes?: string;
  haveYouReportedYesFreeField?: string;
  description?: string;
  agreementForReportingOnBehalf?: string;
  status?: string;
  category?: any[];
  createdAt?: string;
  updatedAt?: string;

  updatereport?: {
    category?: any[];

    description?: string;

    status?: string;
  }[];
};
export type reportType2 = {
  periodOfTime?: string;

  gender?: string;
  age?: string;
  genderFreeField?: string;
  organizationType?: string[];
  numberOfEmployees?: string;
  organizationTypeFreeField?: string;
  happenedOnlineFreeField?: string;
  PeriodOfTimeFreeField?: string;
  locationOnline?: string;
  identity?: string;
  dateRangeState?: any;
  valueDate?: any;

  location?: any;
  _id?: string;

  happenedOverALongPeriodOfTime?: string;

  manifestationOfDiscrimination?: string;
  manifestationOfDiscriminationFreeField?: string;
  otherFormOfDisc?: string;
  otherFormOfDiscFreeField?: string;
  otherFormOfDiscYes?: string;
  haveYouReported?: string;
  haveYouReportedYes?: string;
  haveYouReportedYesFreeField?: string;
  description?: string;
  agreementForReportingOnBehalf?: string;
  status?: string;
  category?: any[];
  createdAt?: string;

  status2?: string;
  category2?: any[];
  updatedAt?: string;

  description2?: string;
};